Clase 7/clase7_lab1/requirements.txt
aws-cdk-lib==2.100.0
boto3==1.34.0
botocore==1.34.0